+++ pkgs/test_core/CHANGELOG.md
@@ -1,8 +1,11 @@
 ## 0.6.19-wip
 
+* Add support for `-c cli` (the native CLI compiler) to the vm platform.
 * Support using the OS platform selector to configure browser tests.
 * Use a DevTools URL instead of a defunct observatory URL.
 * Add flag `--shard-by-file` to control sharding strategy.
+* Allow package_config `3.x.x`.
+
 
 ## 0.6.18